# WIP: static compile part 4 (user-interface)