--- a/base/client.jl
+++ b/base/client.jl
@@ -274,8 +274,8 @@ let reqarg = Set(UTF8String["--home",          "-H",
             end
             # load file immediately on all processors
             if opts.load != C_NULL
-                let file = abspath(bytestring(opts.load))
-                    @everywhere include($file)
+                @sync for p in procs()
+                    @async remotecall_fetch(p, include, bytestring(opts.load))
                 end
             end
             # eval expression