--- a/base/exports.jl
+++ b/base/exports.jl
@@ -1097,7 +1097,6 @@ export
     include,
     include_string,
     reload,
-    compile,
 
 # RTS internals
     finalizer,