--- a/base/interactiveutil.jl
+++ b/base/interactiveutil.jl
@@ -385,7 +385,6 @@ function workspace()
          Expr(:toplevel,
               :(const Base = $(Expr(:quote, b))),
               :(const LastMain = $(Expr(:quote, last)))))
-    empty!(package_list)
     empty!(package_locks)
     nothing
 end