@vtjnash 2015-02-11
Discussion and counter propel proposals are generally always helpful. This is a functional proof-of-concept, I'm less interested in concerns about how it could be intentionally confused, but very interested in whether it is usable with its current interface.