@stevengj 2015-02-11
I think that only an opt-in interface will be usable (because of the many ways that precompilation can be broken), and a reasonable way to opt in would be to require a `function __init__` in the top-level module file.