@timholy 2015-02-11
I'm not sure I fully understand the current interface, but if it requires each user to make specific choices like
```
./julia -J usr/lib/julia/Base.ji --build usr/lib/julia2 <<EOF
using Compat
EOF
```
then I think it's quite a bit too complicated. I like the idea of opt-in, but it should be something that the package author handles for users. I'm less certain that coupling it to `__init__` is a good idea, because there might be times when you'd want an `__init__` but haven't solved all of the issues necessary for full precompilation. What about
```
@precompile module MyPackage
...
end
```