@timholy 2015-02-11
(Sincere apologies if you were just twiddling your thumbs waiting for this kind of feedback. I had the impression that the current interface represented a snapshot of current status rather than something close to "final.")