@vtjnash 2015-02-11
The --build flags exists as a low level interface. The @ cacheable macro also already exists as a simpler method.

The \_\_init\_\_ proposal required a syntactic method definition, so it would be very easy to hide. If precompiling doesn't work, there is no point in providing the method at all amyways.