@vtjnash 2015-02-11
The macro is nice because it doesn't require much in the way of hooks into evaling a module. The method detection method is less visible, but more backwards compatible.