@timholy 2015-02-11
I don't think we care about backward compatibility: once this gets merged, it will change everything. I've basically been keeping a cap on the features in Images.jl because the thought of further increasing load time is unbearable. (There's also the small matter of _writing_ those features...but currently there is a major disincentive to doing so.)

If I were you, I would just go with the `@cacheable` macro.

Anything else?