@IainNZ 2015-02-11
Pardon my ignorance, just want to make sure I understand whats happening is, if we are going with the macro, then:

1. I add `@cacheable` to my package's code, i.e. `@cacheable module IainPkg ... end`
2. Julia, seeing this when a user does `using IainPkg`, looks for a corresponding `.jlc` file for `IainPkg` and uses that if possible (does it hash the `.jl` files or something?) and package loads as quickly as if it was compiled with `userimg.jl` (or close enough?)
3. If it isn't there, it generates the `.jlc` for next time and gives the slow path.