@StefanKarpinski 2015-02-11
Excellent question, @IainNZ. I'm just as ignorant. Having this spelled out in words may help get feedback.