@Mike43110 2015-02-12
More ignorance, what would make a package non-precompilable?