@timholy 2015-02-12
Some packages will require custom handling. For example, any package that makes use of a Dict of types will need to have that Dict rehashed in a [custom `__init__` function](http://docs.julialang.org/en/latest/manual/modules/?highlight=__init__#module-initialization-and-precompilation). Any external C or Fortran library that requires initialization before first usage will need to have that done. If no one has yet set such things up, precompilation would render the package unusuable.