@hayd 2015-02-12
It would be great if there was a `isprecompilable(m::Module)` method to test whether you can do this, and a method to show what must go in your `__init__` function? Or is it the case that neither of these are possible? (Presumably it's not possible as otherwise it'd negates the need for explicitly defining `__init__` and `@precompile` at all.)