@stevengj 2015-02-12
@hayd, there is some discussion of this above—in general, it seems too difficult and fragile to check whether something is precompilable.

Maybe we could have a Lint-like package to check for common precompilability problems, if the following can be detected (excluding `__init__`)

* Global variables initialized to `Dict`, `Set`, or `Ptr` types 
* Any `ccall`s when the module is loaded
* Anything else that might have side effects called when the module is loaded (e.g. file i/o, println, ...)

`@cacheable module Foo ... end` sounds good to me.  But unlike `@timholy` I am worried about backward compatibility with Julia 0.3 (at least); I don't want to suddenly have to start maintaining two branches of every module to use this.  I suppose we could backport `@cacheable` as a no-op to 0.3.