@timholy 2015-02-12
Yes, the no-op `@cacheable` was part of my intention. We don't have to backport it to julia proper if we use `Compat`. Before opening the module, the top level package file should look like this:
```
if VERSION < v"0.4.0-dev"
    using Compat
end

@cacheable module Foo
...
end
```
Since it's the package author inserting `@cacheable`, the same person can insert the `using Compat` logic. That way nothing gets out of sync.