@stevengj 2015-02-13
@timholy, if you do that, won't `using Foo` automatically import `Compat`'s exported symbols into the global namespace?    If so, I'm not sure if that's a good idea.