@timholy 2015-02-13
I suppose if that's a problem, then one could create a `Precompile03.jl` package that contains the one exported symbol `@cacheable`.

An obvious alternative is that package authors define the fallback themselves:
```
if VERSION < v"0.4.0-dev+1234"
    macro cacheable(ex)
        esc(ex)
    end
end

@cacheable module Foo
...
end
```

The final alternative of backporting `@cacheable` to 0.3 seems least attractive, because any package that uses it would break on older 0.3 versions of julia.