@stevengj 2015-02-13
I don't think we want package authors to define it themselves, since if multiple packages do it then you will get warnings, no?  Probably a `CacheableCompat` package is the simplest reasonable thing.