@kmsquire 2015-02-13
> @timholy, if you do that, won't using Foo automatically import Compat's exported symbols into the global namespace? If so, I'm not sure if that's a good idea.

Couldn't you just import the package and qualify the symbol?

```julia
import Compat

Compat.@cachable module Foo
...
end
```