@stevengj 2015-02-13
@kmsquire, that wouldn't work in 0.4 because in that case you want `Base.@cacheable` and not `Compat.@cacheable`.