@rsrock 2015-07-16
Couldn't resist any longer. With this PR:

```
julia> tic(); import Gadfly; toc()
elapsed time: 2.417824882 seconds
2.417824882

julia> @time Gadfly.plot(x=collect(1:100), y=sort(rand(100)))
 708.290 milliseconds (942 k allocations: 39308 KB, 2.83% gc time)
# Actually takes about 10 s after the timing line is returned

julia> @time Gadfly.plot(x=collect(1:100), y=sort(rand(100)))
 130.437 microseconds (79 allocations: 8112 bytes)
# But the second call is nearly instant
```

Before this PR:
```
julia> tic(); import Gadfly; toc()
# depwarns
elapsed time: 28.449467057 seconds
28.449467057

julia> @time Gadfly.plot(x=collect(1:100), y=sort(rand(100)))
 692.214 milliseconds (907 k allocations: 37917 KB, 2.55% gc time)
# About 15 sec elapsed before the plot appears

julia> @time Gadfly.plot(x=collect(1:100), y=sort(rand(100)))
 110.621 microseconds (79 allocations: 8112 bytes)
# This one is instant
```

So the savings is 10x on the `import`